Fix debugger support for direct-launched jobs.
The orte rte component checks the orte_standalone_operation to decide
if it should wait for a message from the hnp or wait on the debugger.
This variable needed to be set to true in ess/pmi to enable the
correct path when direct launching.

cmr=v1.7.3:reviewer=rhc
cmr=v1.6.6:reviewer=rhc

git-svn-id: https://svn.open-mpi.org/svn/ompi/trunk@29013 63e3feb5-37d5-0310-a306-e8a459e722fe