Update the ompi_info man page with information about variable levels and improve the behavior of ompi_info.
This commit changes the default behavior of ompi_info --all when a
level is not specified. Instead of assuming level 1 in this case we
now assume level 9. This change is due to feedback from the community
after the introduction of the --level option.

I also added a new option: --selected-only. This option will limit the
displayed variables to components that can be selected (ie. if there
is a selection parameter set-- btl self,sm)

cmr=v1.7.3:reviewer=jsquyres

git-svn-id: https://svn.open-mpi.org/svn/ompi/trunk@29070 63e3feb5-37d5-0310-a306-e8a459e722fe