Cisco CSCuj13135: Data corruption in  MPI_Bsend_ator_c
Do not assume that the "size" passed to alloc_send() will be the same as
the size of the message the resulting fragment will hold when
usnic_send() is called.  This means usnic_send()/usnic_put() can never
trust any pre-computed size values, and are only allowed to look at the
lengths and pointers of the elements in the desc SG list.



git-svn-id: https://svn.open-mpi.org/svn/ompi/trunk@29183 63e3feb5-37d5-0310-a306-e8a459e722fe